Organize backend modules into package
