+++ src/backend/__init__.py
@@ -1,0 +1,11 @@
+from .backend import ChatBackend
+from .models import Message
+from .settings_manager import AppSettings, load_settings, save_settings
+
+__all__ = [
+    "ChatBackend",
+    "Message",
+    "AppSettings",
+    "load_settings",
+    "save_settings",
+]
