+++ src/backend/settings_manager.py
